requirements.txt
Django==4.0.1
djangorestframework==3.13.1
pdfkit==1.0.0
XlsxWriter==1.4.4
